+++ src/main/java/com/chrisnewland/demofx/util/ShapeEffect.java
@@ -57,10 +57,10 @@ public class ShapeEffect extends AbstractEffect
 
 	private static final int OFFSCREEN = 16;
 
-	private final int OFFSCREEN_LEFT;
-	private final int OFFSCREEN_RIGHT;
-	private final int OFFSCREEN_TOP;
-	private final int OFFSCREEN_BOTTOM;
+	private final double OFFSCREEN_LEFT;
+	private final double OFFSCREEN_RIGHT;
+	private final double OFFSCREEN_TOP;
+	private final double OFFSCREEN_BOTTOM;
 
 	private final PlotMode mode;
 	private final int rotateDegrees;
@@ -103,8 +103,8 @@ public class ShapeEffect extends AbstractEffect
 
 		for (int i = 0; i < itemCount; i++)
 		{
-			int x = getRandomIntInclusive(0, width);
-			int y = getRandomIntInclusive(0, height);
+			int x = getRandomIntInclusive(0, (int)width);
+			int y = getRandomIntInclusive(0, (int)height);
 
 			int dx = getRandomIntInclusive(2, 12);
 			int dy = getRandomIntInclusive(2, 12);
